+++ src/main/java/racing/console/RacingCarConsoleApplication.java
@@ -1,0 +1,12 @@
+package racing.console;
+
+import racing.console.controller.RacingCarConsoleController;
+
+public class RacingCarConsoleApplication {
+
+    public static void main(String[] args) {
+        RacingCarConsoleController racingCarConsoleController = new RacingCarConsoleController();
+
+        racingCarConsoleController.playGame();
+    }
+}
