--- src/main/java/racing/controller/dto/request/validator/RequestValidator.java
@@ -1,6 +1,0 @@
-package racing.controller.dto.request.validator;
-
-public interface RequestValidator<T> {
-
-    void validate(T requestDto);
-}
