+++ src/main/java/racing/domain/Cars.java
@@ -6,12 +6,25 @@ import java.util.stream.Collectors;
 
 public class Cars {
 
+    private static final String CAR_NAME_DUPLICATED = "중복된 자동차 이름이 존재합니다.";
     private static final SecureRandom random = new SecureRandom();
 
     private final List<Car> cars;
 
     public Cars(List<Car> cars) {
+        validateDuplication(cars);
         this.cars = cars;
+    }
+
+    private void validateDuplication(List<Car> cars) {
+        long distinctCount = (int) cars.stream()
+                .map(Car::getName)
+                .distinct()
+                .count();
+
+        if (distinctCount != cars.size()) {
+            throw new IllegalArgumentException(CAR_NAME_DUPLICATED);
+        }
     }
 
     public void moveCarsByCount(int count) {
