+++ src/main/java/racing/web/RacingCarApplication.java
@@ -1,0 +1,13 @@
+package racing.web;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RacingCarApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(RacingCarApplication.class, args);
+    }
+
+}
