+++ src/main/java/racing/web/controller/dto/request/validator/RequestValidator.java
@@ -1,0 +1,6 @@
+package racing.web.controller.dto.request.validator;
+
+public interface RequestValidator<T> {
+
+    void validate(T requestDto);
+}
