--- src/test/java/racing/RacingCarApplicationTests.java
@@ -1,13 +1,0 @@
-package racing;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RacingCarApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
