+++ src/test/java/racing/web/RacingCarApplicationTests.java
@@ -1,0 +1,13 @@
+package racing.web;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RacingCarApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
